Revert "Add two more channels to RC example"
On second thought, because this makes the example incompatible with the Uno I'm going to revert this change. I'm imagining that reading RC receivers will be a common use for this library and I don't want to make one of the more common examples incompatible with the most common board.

This reverts commit c0025f45471d24f6207e7649e438d8641cf0df61.