Remove dependence on PinChangeInterrupt lib
This was well-intentioned but misguided. It prevents the user from having to think about which pin to use, *but* it also prevents advanced users from establishing their own pin change ISR functions for other purposes - whether they're using NicoHood's library for ServoInput or not. It also requires some messy preprocessor checks to see if the board is supported before importing the library, which are inflexible and unlinked to the list of supported boards in that library as well.