Revert "Detach interrupt in destructor"
On second thought, no. Because the interrupt is linked to the static members of the class it's still valid even when a member object is destroyed. More importantly, since the user could make multiple objects of the same pin that thus share the same interrupt vector, this could remove an interrupt that is still in use elsewhere in the program.

This reverts commit 34f719aad613eb29807f9436a3a2d4ff1a92c525.