Add timeout to the loopback() example
Since this is a good test of basic functionality and accuracy, it makes sense to add the new timeout() code to check whether that function works (and to tell the user when the wiring is undone).

I also changed the name of the function and stuck the outputServo call in there (because repeated code makes me sad).