Disable pin checks if no platform support
The library won't compile on Arduino platforms that do not define the `NOT_AN_INTERRUPT` macro. This macro is used by the library to evaluate whether a pin is supported for interrupts.

The previous fix (e7e5845) defines this macro if it's missing, with the assumption that all pins on the platform support interrupts. That didn't solve the issue with the renesas platform (#28) however, as the `digitalPinToInterrupt()` function is a passthrough and not a compile-time constant. That fails the `static_assert()` check.

Instead, this patch adds a bypass. If `NOT_AN_INTERRUPT` is not defined, all code that uses the `digitalPinToInterrupt()` function and the `NOT_AN_INTERRUPT` flag is disabled. There won't be any compile-time or run-time checks for whether the pin is supported or not, but there's nothing we can do about that if the platform doesn't support the feature.