Add flag for interrupt attachment state
The flag indicates whether the interrupt is currently attached, and will prevent the attach/detach functions from calling the platform functions unless it's to change state. This guards against two edge cases.

The first is if the user calls 'attach()' multiple times and the platform does something unexpected when calling the 'attachInterrupt()' function repeatedly. It shouldn't, but you never know and it's something outside of our control. Better to play it safe.

The second, and even more of an edge case, is if the user defines their own interrupt for the pin (not using the library 'attach()' function), and the last class instance is deallocated. This avoids the destructor automatically calling 'detach()' and removing the user's interrupt function. I can't imagine that occuring in practice, but it's a potential pitfall nonetheless.